+++ src/Symfony/Component/Console/DependencyInjection/AddConsoleCommandPass.php
@@ -60,7 +60,7 @@ class AddConsoleCommandPass implements CompilerPassInterface
             $definition->addTag('container.no_preload');
 
             /** @var AsCommand|null $attribute */
-            $attribute = ($r->getAttributes(AsCommand::class)[0] ?? null)?->newInstance();
+            $attribute = ($r->getAttributes(AsCommand::class, \ReflectionAttribute::IS_INSTANCEOF)[0] ?? null)?->newInstance();
 
             if (Command::class !== (new \ReflectionMethod($class, 'getDefaultName'))->class) {
                 trigger_deprecation('symfony/console', '7.3', 'Overriding "Command::getDefaultName()" in "%s" is deprecated and will be removed in Symfony 8.0, use the #[AsCommand] attribute instead.', $class);